Project/e-commerce/requirements.txt
# requirements.txt
pymongo==4.5
# pydantic==2.3
django-ninja-extra==0.19.8
requests==2.31
email-validator==2.0
typing-extensions==4.8
django==4.2
django-bootstrap-v5==1.0.11
Pillow==10.1.0